Move App component into App directory
